Don't use this symlink-based switch anymore
I used it to select different profiles on Nixox vs non-Nixos
computers, but I don't use Nixos anymore. The problem is that it
doesn't play well with Flakes which want to only use nix-checked
files. Having options is still a good idea, but I don't quite know
how to go about it yet.